+++ sshd-core/src/main/java/org/apache/sshd/server/auth/password/PasswordAuthenticator.java
@@ -40,5 +40,23 @@ public interface PasswordAuthenticator {
      * @throws AsyncAuthException If the authentication is performed asynchronously
      */
     boolean authenticate(String username, String password, ServerSession session)
-            throws PasswordChangeRequiredException, AsyncAuthException;
+        throws PasswordChangeRequiredException, AsyncAuthException;
+
+    /**
+     * Invoked when the client sends a {@code SSH_MSG_USERAUTH_REQUEST} indicating
+     * a password change. This can happen if the {@code authenticate} method
+     * threw {@link PasswordChangeRequiredException} thus telling the client that
+     * it needs to provide a new password. Throws {@link UnsupportedOperationException}
+     * by default.
+     *
+     * @param session  The {@link ServerSession} attempting the authentication
+     * @param username The username credential
+     * @param oldPassword The old password
+     * @param newPassword The new password
+     * @return {@code true} if password change accepted
+     */
+    default boolean handleClientPasswordChangeRequest(
+            ServerSession session, String username, String oldPassword, String newPassword) {
+        throw new UnsupportedOperationException("Password change not supported");
+    }
 }