+++ sshd-core/src/test/java/org/apache/sshd/util/test/BogusPasswordAuthenticator.java
@@ -23,7 +23,8 @@ import org.apache.sshd.server.auth.password.PasswordAuthenticator;
 import org.apache.sshd.server.session.ServerSession;
 
 /**
- * TODO Add javadoc
+ * A test {@link PasswordAuthenticator} that accepts an authentication
+ * attempt if the username is not {@code null} and same as password
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
@@ -38,7 +39,8 @@ public class BogusPasswordAuthenticator extends AbstractLoggingBean implements P
     public boolean authenticate(String username, String password, ServerSession session) {
         boolean result = (username != null) && username.equals(password);
         if (log.isDebugEnabled()) {
-            log.debug("authenticate({}) {} / {} - success = {}", session, username, password, result);
+            log.debug("authenticate({}) {} / {} - success={}",
+                session, username, password, result);
         }
 
         return result;